Alpha 0.1
First working version.
Threshold: threshold of binary conversion
Brightness: Brightness of original image
Contrast: Contrast of original image
MinSqSize: Minimum size of squares to detect
MaxCosine: Max skew of squares, for detecting squares in perspective
(!=90 degrees)
KSize: kernel size of the dilate function (unsharpens original image,
to filter out noise)
CannyThreshold: Threshold for the Canny function, makes lines «come
together»
Delay: Delay from a square is lost in the image until it is updated as
gone, to filter out hand movements above the squares/on the squares
Distance: Sets the minimum distance between squares. If they are closer
than this, they will be detected with the same placement.

OUTPUT:
there are four numbers updated in the console:
The first is the number of the square (0-19)
The second is the x coordinate
The third is the y coordinate
The fourth is the angle of the square

If there are more than 20 squares detected, the console will display a
warning to adjust the filters.
If there are no squares, the condole will say nothing.

These numbers are also sent via UDP (OSC) to the selected IP
address(es)